+++ src/main/java/nl/syntouch/dmn/studio/model/dto/DMNVersionDTO.java
@@ -2,7 +2,7 @@ package nl.syntouch.dmn.studio.model.dto;
 
 public record DMNVersionDTO(
         Long id,
-        Integer version,
+        Long version,
         byte[] fileBlob,
         Integer status
 ) {
